requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
opencv-python==4.10.0.84
bokeh==3.3.4
bottle==0.12.25
branca==0.7.1
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
clr-loader==0.2.6
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
descartes==1.1.0
executing==2.0.1
fiona==1.9.6
Flask==3.0.2
flatbuffers==24.3.25
folium==0.16.0
fonttools==4.49.0
gast==0.5.4
geopandas==0.14.3
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.11.0
idna==2.10
importlib_metadata==7.0.2
ipykernel==6.29.4
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.2.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mplcursors==0.5.3
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
prompt-toolkit==3.0.43
protobuf==4.25.3
proxy_tools==0.1.0
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.2
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.2
pythonnet==3.0.3
pytz==2024.1
pywebview==5.0.5
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.0
requests==2.25.1
retrying==1.3.4
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
setuptools==69.1.1
shapely==2.0.3
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.3.0
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.6
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
xyzservices==2023.10.1
zipp==3.17.0
